requirements.txt
openai>=1.30.0
anthropic>=0.25.0
python-dotenv>=1.0.0
Pillow>=10.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
sendgrid>=6.10.0
stripe>=7.4.0
fpdf2>=2.7.0
weasyprint>=59.0
playwright>=1.40.0
playwright-stealth>=1.0.1
imagehash>=4.3.0
httpx>=0.25.0
